relocates feature card
